indexer/requirements.txt
# ==============================================================================
# Python Indexer Service - Dependencies
# ==============================================================================

# FastAPI framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Milvus vector database
pymilvus==2.4.0

# NumPy - MUST be 1.x for compatibility with torch and sentence-transformers
numpy<2.0.0

# Embedding model - pin compatible versions
torch==2.1.2
transformers==4.36.2
sentence-transformers==2.3.1

# PDF processing
PyMuPDF==1.23.8
pdf2image==1.16.3

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Text processing
langdetect==1.0.9
unidecode==1.3.8

# File watching
watchdog==3.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Logging
loguru==0.7.2


# Fix marshmallow version for pymilvus compatibility
marshmallow>=3.13.0,<3.22.0
environs>=9.0.0,<12.0.0
